Restructure to separate main test.
