+++ workflow_scheduler_test.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"github.com/microsoft/BladeMonRT/nodes"
-	"github.com/microsoft/BladeMonRT/nodes/dummy_node_a"
 	//"github.com/microsoft/BladeMonRT/nodes/dummy_node_b"
 	workflows "github.com/microsoft/BladeMonRT/workflows"
 	gomock "github.com/golang/mock/gomock"
@@ -14,9 +12,6 @@ import (
 	"log"
 	"github.com/microsoft/BladeMonRT/logging"
 	"time"
-	"io/ioutil"
-	"github.com/microsoft/BladeMonRT/test_configs"
-	"gotest.tools/assert"
 )
 
 type UtilsForTest struct {
@@ -24,40 +19,6 @@ type UtilsForTest struct {
 
 func (utilsForTest UtilsForTest) ParseEventXML(eventXML string) utils.EtwEvent {
 	return utils.EtwEvent{}
-}
-
-func TestSetupWorkflowScheduler(t *testing.T) {
-	// Assume
-	workflowsJson, err := ioutil.ReadFile(test_configs.TEST_WORKFLOW_FILE)
-	if err != nil {
-		log.Fatal(err)
-	}
-	var workflowFactory WorkflowFactory = newWorkflowFactory(workflowsJson, NodeFactory{})
-
-	schedulesJson, err := ioutil.ReadFile(test_configs.TEST_SINGLE_SCHEDULE_FILE)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	var mainObj *Main = newMain()
-	mainObj.setupWorkflows(schedulesJson, workflowFactory)
-
-	// Action
-	var workflowScheduler *WorkflowScheduler = mainObj.WorkflowScheduler
-
-	// Assert
-	// Assert that the GUID to context map has only 1 context.
-	assert.Equal(t, len(workflowScheduler.guidToContext), 1)
-	
-	// Check that the first and second nodes in the context's workflow are of type DummyNodeA by checking the value of result field in the objects.
-	var firstWorkflowNodes []nodes.InterfaceNode
-	for _, element := range workflowScheduler.guidToContext {
-		firstWorkflowNodes = element.workflow.GetNodes()
-    }
-	actualFirstNode :=  *(firstWorkflowNodes[0]).(*dummy_node_a.DummyNodeA)
-	actualSecondNode := *(firstWorkflowNodes[1]).(*dummy_node_a.DummyNodeA)
-	assert.Equal(t, actualFirstNode.Result, "dummy-node-result")
-	assert.Equal(t, actualSecondNode.Result, "dummy-node-result")
 }
 
 func TestSubscriptionCallback(t *testing.T) {
